+++ client-app/lib/views/settings_screen.dart
@@ -4,8 +4,8 @@ import 'package:get/get.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:plan_sync/util/constants.dart';
 import 'package:plan_sync/util/external_links.dart';
-import 'package:plan_sync/widgets/bottom_sheets.dart';
-import 'package:plan_sync/widgets/logout_button.dart';
+import 'package:plan_sync/widgets/bottom-sheets/bottom_sheets_wrapper.dart';
+import 'package:plan_sync/widgets/buttons/logout_button.dart';
 import '../controllers/auth.dart';
 
 class SettingsPage extends StatelessWidget {
