+++ client-app/lib/widgets/buttons/logout_button.dart
