+++ client-app/lib/widgets/buttons/schedule_preferences_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:plan_sync/controllers/filter_controller.dart';
-import 'package:plan_sync/widgets/bottom_sheets.dart';
+import 'package:plan_sync/widgets/bottom-sheets/bottom_sheets_wrapper.dart';
 
 /// Usually on the home screen used to select
 /// required time table.
