+++ client-app/lib/widgets/dropdowns/electives_sem_bar.dart
