+++ client-app/lib/widgets/dropdowns/semester_bar.dart
